Move related script to one folder with Header
