+++ packages/frontend/src/components/index.ts
@@ -1,7 +1,7 @@
 export * from './chart/Chart'
 export * from './Footer'
 export * from './head'
-export * from './Header'
+export * from './header/Header'
 export * from './Heading'
 export * from './Logo'
 export * from './navbar/Navbar'
