+++ packages/frontend/src/scripts/index.ts
@@ -1,9 +1,9 @@
 import { configureChart } from '../components/chart/configure'
+import { configureTvlActivity } from '../components/header/configureTvlActivity'
 import { configureDarkThemeToggle } from '../components/navbar/configureDarkThemeToggle'
 import { configureSidebarMenu } from '../components/navbar/configureSidebarMenu'
 import { configureCombinedBridges } from './configureCombinedBridges'
 import { configureTooltips } from './configureTooltips'
-import { configureTvlActivity } from './configureTvlActivity'
 
 configureCombinedBridges()
 configureDarkThemeToggle()
